services/dicpic/requirements.txt
fastapi
uvicorn
pydantic
pyjwt
python-dotenv
python-multipart
psycopg2-binary